backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mutagen==1.47.0
pydantic==2.5.0
httpx==0.25.0
spotipy==2.23.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
